fixed memory leaks

+++ source/gui/blob.h
@@ -3,9 +3,9 @@
 
 #include "quickjs.h"
 
-int JS_GetBlobSize(JSContext *ctx, JSValueConst blob);
+size_t JS_GetBlobSize(JSContext *ctx, JSValueConst blob);
 JSValue JS_GetBlobParts(JSContext *ctx, JSValueConst blob_obj);
 
-int JS_GetBlobUint8Array(JSContext *ctx, JSValueConst blob, uint8_t *out_buf);
+size_t JS_GetBlobUint8Array(JSContext *ctx, JSValueConst blob, uint8_t *out_buf);
 
 #endif
